move new bits to new module names
